refactor(capture): extract WebSocket connection tracking into internal/capture/wsconn
The reference graph (not the filenames) picks this boundary out: ws_connection_store.go,
ws_connection_tracker.go and websocket_status.go form a closed cluster whose only edges
into the rest of capture are four method calls on c.wsConnections. websocket_status.go is
named with the WS *event buffer* files but every symbol in it belongs to the *connection
tracker*; removeFromSlice sat in helpers.go ("shared capture helpers") with exactly one
consumer, this cluster.

wsconn imports only internal/types and internal/util — no capture import, no cycle. The
"caller owns the mutex" contract the type already had is now stated explicitly in doc.go
instead of being implied by struct placement.

capture's exported symbol set is unchanged: WSConnectionTracker is now a type alias for
wsconn.Tracker (go doc -all diff shows only doc-comment movement).

Tests move with the code: ws_connection_store_test.go, the removeFromSlice tests from
helpers_test.go, and the formatAge/formatDuration tests from websocket_test.go. Six tests
were added for functions that would otherwise read 0.0% in the new package's own profile
(they were covered cross-package before); each was verified RED against a mutation of the
code it covers.

internal/capture: 67 -> 65 non-test files. New folder internal/capture/wsconn: 3.
