+++ internal/capture/accessor_counts_health.go
@@ -109,7 +109,7 @@ func (c *Capture) GetHealthSnapshot() HealthSnapshot {
 		WebSocketCount:        c.buffers.webSocketCount(),
 		NetworkBodyCount:      c.buffers.networkCount(),
 		ActionCount:           c.buffers.actionCount(),
-		ConnectionCount:       c.wsConnections.connectionCount(),
+		ConnectionCount:       c.wsConnections.Count(),
 		LastPollTime:          c.extensionState.lastPollAt,
 		ExtSessionID:          c.extensionState.extSessionID,
 		ExtSessionChangedTime: c.extensionState.extSessionChangedAt,
