+++ internal/capture/buffer-types.go
@@ -16,11 +16,3 @@ type NetworkWaterfallBuffer struct {
 type ExtensionLogBuffer struct {
 	logs []ExtensionLog // Ring buffer of extension internal logs (max MaxExtensionLogs=500)
 }
-
-// WSConnectionTracker groups WebSocket connection tracking fields.
-// Protected by parent Capture.mu (no separate lock).
-type WSConnectionTracker struct {
-	connections map[string]*connectionState // Active WS connections by ID (max 20 total). LRU eviction via connOrder.
-	closedConns []WebSocketClosedConnection // Ring buffer of closed connections (max 10, maxClosedConns). Preserves history for a while.
-	connOrder   []string                    // Insertion order for LRU eviction of active connections.
-}
